modelExecutor/requirements.txt
boto3
httpx
pandas
nltk
boto3-stubs[sqs]
google-generativeai